Refactor and rename of the rule
