+++ rules-tests/TypeDeclaration/Rector/FunctionLike/AddClosureParamTypeFromArgRector/Fixture/fixture.php.inc
@@ -1,0 +1,25 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddClosureParamTypeFromArgRector\Fixture;
+
+use SomeNamespace\SomeClass;
+use SomeNamespace\AnotherClass;
+
+SomeClass::someCall('test', function ($object) {
+    return $object;
+});
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddClosureParamTypeFromArgRector\Fixture;
+
+use SomeNamespace\SomeClass;
+use SomeNamespace\AnotherClass;
+
+SomeClass::someCall('test', function (string $object) {
+    return $object;
+});
+
+?>
