+++ rules-tests/TypeDeclaration/Rector/FunctionLike/AddClosureParamTypeFromArgRector/Fixture/skip_if_calllike_arg_is_named.php.inc
@@ -1,0 +1,9 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddClosureParamTypeFromArgRector\Fixture;
+
+use Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddClosureParamTypeFromArgRector\Source\ClassForNamedParameters;
+
+SomeClass::someCall(anotherCallback: fn ($var) => $var, callback: fn($var) => $var);
+
+?>
