+++ rules-tests/TypeDeclaration/Rector/FunctionLike/AddClosureParamTypeFromArgRector/Fixture/skip_if_non_functionlike_parameters_in_method_call.php.inc
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddClosureParamTypeFromArgRector\Fixture;
+
+use SomeNamespace\SomeClass;
+
+SomeClass::someCall('test');
