+++ rules-tests/TypeDeclaration/Rector/FunctionLike/AddClosureParamTypeFromArgRector/Fixture/skip_if_non_matching_class.php.inc
@@ -1,0 +1,9 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddClosureParamTypeFromArgRector\Fixture;
+
+use SomeNamespace\SomeClass;
+
+AnotherClass::someCall(function ($name) {
+    return $name;
+});
