--- rules-tests/TypeDeclaration/Rector/FunctionLike/AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector/Fixture/overrides_previous_type.php.inc
@@ -1,19 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector\Fixture;
-
-use SomeNamespace\SomeClass;
-
-SomeClass::someOtherCall('test', fn(int|string $var) => $var);
-
-?>
------
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector\Fixture;
-
-use SomeNamespace\SomeClass;
-
-SomeClass::someOtherCall('test', fn(string $var) => $var);
-
-?>
