--- rules-tests/TypeDeclaration/Rector/FunctionLike/AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector/Fixture/skip_if_non_functionlike_parameters_in_method_call.php.inc
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector\Fixture;
-
-use SomeNamespace\SomeClass;
-
-SomeClass::someCall('test');
