--- rules-tests/TypeDeclaration/Rector/FunctionLike/AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector/Fixture/skip_if_non_matching_class.php.inc
@@ -1,9 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector\Fixture;
-
-use SomeNamespace\SomeClass;
-
-AnotherClass::someCall(function ($name) {
-    return $name;
-});
