--- rules-tests/TypeDeclaration/Rector/FunctionLike/AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector/Fixture/skip_on_class_string_only.php.inc
@@ -1,12 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\FunctionLike\AddParamTypeForFunctionLikeWithinCallLikeArgFromArgDeclarationRector\Fixture;
-
-use SomeNamespace\SomeClass;
-use SomeNamespace\AnotherClass;
-
-SomeClass::someCall('test', function ($object) {
-    return $object;
-});
-
-?>
